naively move consitency checker call into collection ownership type factory
